requirements.txt
asgiref==3.3.1
Django==3.1.6
pkg-resources==0.0.0
pytz==2021.1
sqlparse==0.4.1
djangorestframework
markdown
django-filter
opencv-python
pillow

# Django REST Framework
djangorestframework  # https://github.com/encode/django-rest-framework
coreapi  # https://github.com/core-api/python-client

# Django Webpack Loader
django-webpack-loader  # https://github.com/owais/django-webpack-loader 

# Headers
django-cors-headers
